+++ src/Plugin.php
@@ -6,7 +6,6 @@ use benf\neo\controllers\Configurator as ConfiguratorController;
 use benf\neo\controllers\Conversion as ConversionController;
 use benf\neo\controllers\Input as InputController;
 use benf\neo\elements\Block;
-use benf\neo\enums\NewBlockMenuStyle;
 use benf\neo\fieldlayoutelements\ChildBlocksUiElement;
 use benf\neo\gql\interfaces\elements\Block as NeoGqlInterface;
 use benf\neo\integrations\feedme\Field as FeedMeField;
@@ -22,13 +21,11 @@ use craft\base\Plugin as BasePlugin;
 use craft\console\Application as ConsoleApplication;
 use craft\console\Controller;
 use craft\console\controllers\ResaveController;
-use craft\controllers\ElementsController;
 use craft\db\Query;
 use craft\db\Table;
 use craft\elements\conditions\SlugConditionRule;
 use craft\elements\GlobalSet;
 use craft\events\DefineConsoleActionsEvent;
-use craft\events\DefineElementEditorHtmlEvent;
 use craft\events\DefineFieldLayoutElementsEvent;
 use craft\events\RebuildConfigEvent;
 use craft\events\RegisterComponentTypesEvent;
@@ -42,7 +39,6 @@ use craft\gatsbyhelper\events\RegisterIgnoredTypesEvent;
 use craft\gatsbyhelper\services\Deltas;
 use craft\helpers\Console;
 use craft\helpers\Db;
-use craft\helpers\Html;
 use craft\models\FieldLayout;
 use craft\services\Fields;
 use craft\services\Gc;
@@ -126,7 +122,6 @@ class Plugin extends BasePlugin
         $this->_registerGatsbyHelper();
         $this->_registerFeedMeSupport();
         $this->_registerConditionFieldRuleRemoval();
-        $this->_registerDefaultBlockTypeIcon();
     }
 
     /**
@@ -373,27 +368,6 @@ class Plugin extends BasePlugin
                             return true;
                         }
                     );
-                }
-            }
-        );
-    }
-
-    private function _registerDefaultBlockTypeIcon()
-    {
-        Event::on(
-            ElementsController::class,
-            ElementsController::EVENT_DEFINE_EDITOR_CONTENT,
-            function(DefineElementEditorHtmlEvent $event) {
-                if ($this->getSettings()->newBlockMenuStyle !== NewBlockMenuStyle::Classic && !$event->static) {
-                    $svg = Html::tag(
-                        'div',
-                        Html::modifyTagAttributes(
-                            Html::svg('@benf/neo/resources/default-new-block-icon.svg'),
-                            ['id' => 'ni-icon'],
-                        ),
-                        ['class' => 'hidden'],
-                    );
-                    $event->html = $svg . $event->html;
                 }
             }
         );